fix crash issues when number more than 11 digit

+++ domain/src/main/java/com/pens/planduit/domain/models/entity/BudgetResult.kt
@@ -1,8 +1,8 @@
 package com.pens.planduit.domain.models.entity
 
 data class BudgetResult(
-    val income: Int,
-    val needs: Int,
-    val wants: Int,
-    val savings: Int
+    val income: Long,
+    val needs: Long,
+    val wants: Long,
+    val savings: Long
 )
