+++ domain/src/main/java/com/pens/planduit/domain/models/request/AgricultureZakatRequest.kt
@@ -3,7 +3,7 @@ package com.pens.planduit.domain.models.request
 import com.google.gson.annotations.SerializedName
 
 data class AgricultureZakatRequest(
-    @SerializedName("total_harvest") val totalHarvest : Int,
+    @SerializedName("total_harvest") val totalHarvest : Long,
     @SerializedName("is_watered") val isWatered : Boolean,
-    @SerializedName("grain_price") val grainPrice : Int
+    @SerializedName("grain_price") val grainPrice : Long
 )
