+++ domain/src/main/java/com/pens/planduit/domain/models/request/BudgetingRequest.kt
@@ -1,5 +1,5 @@
 package com.pens.planduit.domain.models.request
 
 data class BudgetingRequest(
-    val income : Int,
+    val income : Long,
 )
