+++ domain/src/main/java/com/pens/planduit/domain/models/request/SavingZakatRequest.kt
@@ -1,7 +1,7 @@
 package com.pens.planduit.domain.models.request
 
 data class SavingZakatRequest(
-    val savings: Int,
+    val savings: Long,
     val bank: Boolean,
     val interest : Int
 )
